feat: merge cpp_io_stream.h into c2pa.hpp The CppIOStream classes are now public and can be reused if needed. deletes cpp_io_stream.h
